requirements.txt
Django>=4.1.7,<=4.2
djangorestframework>=3.14.0,<=3.15
mysqlclient>=2.1.1,<=2.2
drf-spectacular>=0.26.0,<=0.27
